16: Created config manager singleton
