+++ Scenes/Components/Modal/Settings/Modal.cs
@@ -1,27 +1,26 @@
 using Godot;
 using System;
 using System.Collections.Generic;
+using GodotPeer2PeerSteamCSharp.Modules.Config.Video;
 using GodotPeer2PeerSteamCSharp.Scenes.Components.Modal.Settings;
 
 public partial class Modal : Panel
 {
     private Button _confirmButton;
     
-    private VideoSettings _videoSettings;
+    private VideoTab _VideoTab;
 
     public override void _Ready()
     {
         _confirmButton = GetNode<Button>("%Confirm");
         _confirmButton.Pressed += SaveSettings;
         
-        _videoSettings = GetNode<VideoSettings>("%VideoSettings");
+        _VideoTab = GetNode<VideoTab>("%VideoTab");
     }
 
     private void SaveSettings()
     {
         var config = new ConfigFile();
-        
-        _videoSettings.SaveSettings(config);
         
         Error errorResult = config.Save("user://settings.cfg");
