+++ Scenes/Components/Modal/Settings/Modal.tscn
@@ -1,7 +1,7 @@
 [gd_scene load_steps=5 format=3 uid="uid://c2w2wj8r5dic8"]
 
 [ext_resource type="Script" path="res://Scenes/Components/Modal/Settings/Modal.cs" id="1_qm5fi"]
-[ext_resource type="PackedScene" uid="uid://cjv0qy65wg5y3" path="res://Scenes/Components/Modal/Settings/VideoSettings.tscn" id="2_wx8pc"]
+[ext_resource type="PackedScene" uid="uid://cjv0qy65wg5y3" path="res://Scenes/Components/Modal/Settings/Video/VideoTab.tscn" id="2_4xncj"]
 
 [sub_resource type="StyleBoxFlat" id="StyleBoxFlat_54awh"]
 bg_color = Color(0.151276, 0.151276, 0.151276, 1)
@@ -47,8 +47,7 @@ grow_vertical = 2
 text = "Settings"
 label_settings = SubResource("LabelSettings_uu3k7")
 
-[node name="VideoSettings" parent="VBoxContainer" instance=ExtResource("2_wx8pc")]
-unique_name_in_owner = true
+[node name="VideoTab" parent="VBoxContainer" instance=ExtResource("2_4xncj")]
 layout_mode = 2
 
 [node name="HBoxContainer" type="HBoxContainer" parent="VBoxContainer"]
