+++ project.godot
@@ -20,6 +20,7 @@ config/icon="res://icon.svg"
 LobbyManager="*res://Modules/Lobby/LobbyManager.cs"
 TransportManager="*res://Modules/Transport/TransportManager.cs"
 SceneManager="*res://Modules/Scene/SceneManager.cs"
+ConfigManager="*res://Modules/Config/ConfigManager.cs"
 UIManager="*res://Modules/UI/UIManager.cs"
 BackgroundManager="*res://Modules/Background/BackgroundManager.cs"
 InputManager="*res://Modules/Input/InputManager.cs"
